requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blas==1.0
boost==1.74.0
boost-cpp==1.74.0
brotli-python==1.0.9
bzip2==1.0.8
ca-certificates==2024.2.2
cairo==1.16.0 
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
cpuonly==2.0
cycler==0.12.1
datasets==3.0.1
descriptastorus==2.6.1
dill==0.3.8
expat==2.6.2
ffmpeg==4.3
filelock==3.13.1 
fontconfig==2.14.1
freetype==2.10.4
frozenlist==1.4.1
fsspec==2023.10.0
glib==2.80.0
glib-tools==2.80.0
gmp==6.2.1
gmpy2==2.1.2
gnutls==3.6.15
greenlet==3.0.3
grpcio==1.67.0
huggingface-hub==0.26.0
icu==70.1
idna==3.4
jinja2==3.1.3
joblib==1.3.2
jpeg==9e
kiwisolver==1.4.5
lame==3.100
lcms2==2.12
ld_impl_linux-64==2.38
lerc==3.0
libblas==3.9.0
libcblas==3.9.0
libdeflate==1.17
libexpat==2.6.2
libffi==3.4.4
libgcc-ng==13.2.0
libgfortran-ng==13.2.0
libgfortran5==13.2.0
libglib==2.80.0
libiconv==1.17
libidn2==2.3.4
libjpeg-turbo==2.0.0
liblapack==3.9.0
libopenblas==0.3.26
libpng==1.6.43
libprotobuf==3.20.3
libstdcxx-ng==13.2.0
libtasn1==4.19.0
libtiff==4.5.1
libunistring==0.9.10
libuuid==1.41.5
libwebp-base==1.3.2
libxcb==1.16
libxml2==2.9.14
libzlib==1.2.13
llvm-openmp==18.1.1
markdown==3.7
markupsafe==2.1.3
matplotlib-base==3.4.3
mock==5.1.0
mpc==1.1.0
mpfr==4.0.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
ncurses==6.4
nettle==3.7.3
networkx==3.2.1
ninja==1.10.2
ninja-base==1.10.2
numpy==1.26.4
openh264==2.1.1
openjpeg==2.4.0
openssl==3.2.1
packaging==24.0
pandas==2.2.1
pandas-flavor==0.6.0
pcre2==10.43
pillow==10.2.0
pip==23.3.1
pixman==0.43.2
protobuf==5.28.2
psutil==5.9.8
pthread-stubs==0.4
pyarrow==17.0.0
pycairo==1.23.0
pycparser==2.21
pyparsing==3.1.2
pysocks==1.7.1
python==3.10.13
python-dateutil==2.9.0
python-tzdata==2024.1
python_abi==3.10
pytorch==2.2.0
pytorch-mutex==1.0
pytz==2024.1
pyyaml==6.0.2
rdkit==2023.9.5
readline==8.2
reportlab==3.6.13
requests==2.32.3
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==68.2.2
six==1.16.0
sqlalchemy==2.0.28
sqlite==3.41.2
subword-nmt==0.3.8
sympy==1.12
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.3.0
tk==8.6.12
torch-geometric==2.5.1
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.5
typing-extensions==4.10.0
typing_extensions==4.10.0
tzdata==2024a
urllib3==2.1.0
werkzeug==3.0.4
wheel==0.41.2
xarray==2024.2.0
xgboost==2.0.3
xorg-libxau==1.0.11
xorg-libxdmcp==1.1.3
xxhash==3.5.0
xz==5.4.6
yarl==1.9.4
zlib==1.2.13
zstd==1.5.5